Rearrange and rename registrants folders, files, and modules
