+++ course-management-tools/registrants_main.py
@@ -1,12 +1,12 @@
 
 from typing import NamedTuple
 from unittest.mock import Mock
-from src.registration.core.workflows import RegistrationWorkflows
-from src.registration.adapters.registrants_repo_inmemory import InMemoryRegistrantsRepo
+from src.registrants.core.workflows import RegistrantsWorkflows
+from src.registrants.adapters.registrants_repo_inmemory import InMemoryRegistrantsRepo
 from src.external.console import Console
-from src.registration.adapters.contact_info_formatter_gmail import GmailContactInfoFormatter
-from src.registration.adapters.contact_info_presenter_print import PrintContactInfoPresenter
-from src.cli.registrants_cli_test import RegistrantsCLI
+from src.registrants.adapters.contact_info_formatter_gmail import GmailContactInfoFormatter
+from src.registrants.adapters.contact_info_presenter_print import PrintContactInfoPresenter
+from src.registrants.interactors.cli import RegistrantsCLI
 
 class MockRegistrant(NamedTuple):
     name: str
@@ -61,7 +61,7 @@ cli.get_input.return_value = "workshop2"
 
 interactor = RegistrantsCLI(
     cli=cli, 
-    workflows=RegistrationWorkflows(
+    workflows=RegistrantsWorkflows(
         registrants_repo=InMemoryRegistrantsRepo(workshops),
     ), 
     presenter=PrintContactInfoPresenter(
