+++ course-management-tools/src/registrants/adapters/contact_info_formatter_gmail.py
