+++ course-management-tools/src/registrants/adapters/registrants_repo_inmemory.py
