+++ course-management-tools/src/registrants/core/contact_info_formatter.py
