+++ course-management-tools/src/registrants/core/entities.py
